Changed AbstractNumberSetting's MAX_VALUE and MIN_VALUE to be Numbers instead of Objects.

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/settings/ByteSetting.java
@@ -60,8 +60,8 @@ public final class ByteSetting extends AbstractNumberSetting {
     }
 
     protected boolean isAcceptableValue(String value) {
-        byte max = ((Byte)MAX_VALUE).byteValue();
-        byte min = ((Byte)MIN_VALUE).byteValue();
+        byte max = MAX_VALUE.byteValue();
+        byte min = MIN_VALUE.byteValue();
         byte val = Byte.parseByte(value);
         return (val <= max && val >= min);
     }
